+++ src/Ext/Techno/Hooks.cpp
@@ -642,3 +642,16 @@ DEFINE_HOOK(0x70265F, TechnoClass_ReceiveDamage_Explodes, 0x6)
 
 	return 0;
 }
+
+DEFINE_HOOK(0x703A09, TechnoClass_VisualCharacter_ObserverCloak, 0x6)
+{
+	enum { UseShadowyVisual = 0x703A5A };
+
+	GET(TechnoClass*, pThis, ESI);
+
+	// Allow observers to always see cloaked objects.
+	if (HouseClass::IsCurrentPlayerObserver() && pThis->CloakState == CloakState::Cloaked)
+		return UseShadowyVisual;
+
+	return 0;
+}